Add device activation for screen recording
Before we can start the screen capture protocol, we first need to send a USB control message in order for the device to activate the hidden USB configuration over which recording data is transmitted. This operation causes a disconnection briefly, so reconnect with some retries.

Deactivation is left for future change.